Move image compare functions to imgcompare file
