+++ apps/rendering/task/verificator.py
@@ -12,8 +12,7 @@ from golem.task.taskbase import ComputeTaskDef
 from golem.task.localcomputer import LocalComputer
 
 from apps.core.task.verificator import CoreVerificator, SubtaskVerificationState
-from apps.rendering.resources.imgcompare import check_size
-from apps.rendering.resources.imgrepr import advance_verify_img
+from apps.rendering.resources.imgcompare import advance_verify_img, check_size
 
 
 logger = logging.getLogger("apps.rendering")
