+++ tests/apps/rendering/resources/test_renderingtaskcollector.py
@@ -6,7 +6,8 @@ from golem.tools.testdirfixture import TestDirFixture
 
 
 from apps.rendering.resources.renderingtaskcollector import RenderingTaskCollector
-from apps.rendering.resources.imgrepr import compare_pil_imgs, advance_verify_img, load_img
+from apps.rendering.resources.imgcompare import compare_pil_imgs, advance_verify_img
+from apps.rendering.resources.imgrepr import load_img
 
 
 def make_test_img(img_path, size=(10, 10), color=(255, 0, 0)):
